requirements.txt
alembic==1.7.6
anyio==3.5.0
asgiref==3.4.1
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
decorator==5.1.1
executing==0.8.2
fastapi==0.71.0
FastAPI-SQLAlchemy==0.2.1
greenlet==1.1.2
h11==0.12.0
idna==3.3
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.0.1
jedi==0.18.1
Mako==1.1.6
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
pluggy==1.0.0
prompt-toolkit==3.0.26
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pydantic==1.9.0
Pygments==2.11.2
pyparsing==3.0.6
pytest==6.2.5
python-dotenv==0.19.2
requests==2.27.1
schedule==1.1.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.29
stack-data==0.1.4
starlette==0.17.1
toml==0.10.2
tomli==2.0.0
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.0
wcwidth==0.2.5
